READ ONLY: functional simulation
WARNING: This commit is intended for viewing purposes only.  Subsequent commits will implement significant changes to the structure of the simulation. Consequently, this commit should not be used as a base for future developmental branches. 

WARNING: Due to the nature of subsequent commits as noted above, each commit would require a separate readme file.  Because of this, the current readme file is NOT up to date, and the author provides brief instructions on running the simulation bellow.

Required changes for each user:
In IonWake_000.cu change dataDirName to the absolute path name of the directory you want the simulation output to be stored.
In IonWake_000.cu change inputDirName to the absolute path name of the directory you want the simulation to read impute files from.

To compile:
	All the following is intended for use with KODIAK
	Place all .h .cpp and .cu files into the same directory as the included makefile.
	Run the command “module load cuda80/toolki