Add files via upload
Added cylindrical boundary conditions.
The incoming ions have velocities that are consistent with incidence on the top, side, or bottom of the cylinder.  The electric field due to ions outside of the cylinder is calculated using a Matlab code, e_field_in_cylinder.m, and then fit with a high-order polynomial for fast calculation within the code.
In the params.txt file, cylindrical boundary conditions are specified with a 1 (a '0' for geometry designates spherical boundary conditions).  The number of divisions used in the incoming angle (NUM_QTH_DIV) is set to 3, corresponding to top, side, bottom.  
The coefficients for the polynomial fits for the electric field are also input in params.txt (P10X - P05Z).