Created checkCudaError_000 Function
The checkCudaError_000 function has been added to the
IonWake_000.cu file. The function is used to check for CUDA errors
and will write an error mesage to stderr and terminate the program
if an error is present. The function will eventualy need to be moved
into a seperate file and renamed with the apropriate number ending.