Replace Error Handeling With checkCudaError_000
Replace all instances of CUDA error handeling in IonWake_000.cu
with the checkCudaError_000() function. The function does not
change any of the error handeling functionality but just replaces
a repeditve section of code.

The error messages for all of the checkCudaError_000()
function calls are currently the same. This will be
fixed in future commits.