checkCudaError_000 -> roadBlock_000
Changed CUDA error handeling in IonWake_000.cu from
checkCudaError_000() to roadBlock_000.

roadBlock_000 can print a string to the status file. It then
checks for a CUDA kernel launch error and waits for the
threads to synchronize. Finaly it checks if cudaDeviceSynchorinze
returns any errors. If it finds an error at any point it prints
an error to standard error and terminates the program.