Added Parameter: FREQ
Added the parameter FREQ to IonWake_000.cu and params.txt. FREQ
gives the frequency of the applied electric feild and can be used
to creat a symetric ion wakefield.

Re-wrote the calculation of xac in the timestep. It apears that
floor() % 2 was attempted. However, floor() returns a floating
point value and % requires two int inputs. This has been fixed with
int(floor()) % 2.