Added Timestep Notes
Notes were added to IonWake_000.cu in the time step to denote
non-constant input and ouput parameters for select_100(). In
future commits, notes will be added to most time step function
calls. "// -->" denotes an output and "// <--" denotes and input.
"// ?? text ??" denotes a question or concern regarding a
constant or non-constant parameter. These notes are intended to
aid in the tracking of variables and memory in the timestep and
not to be a permenant fixture.

In the IonWake_000.cu time step a single line calculation was
split into multiple lines.