Added routine to calc ion number density and potential
calcIonDensity_102 was added to IonWake_102*.  In the main program,
grid points are determined at which to find the ion density and the
potential due to the ions.  The number of grid points is a multiple
of BLOCK_DIM.  The ion density is calculated by finding the number
of ions within a radius of DEBYE/5 and divided by the volume of this
test sphere.  The grid points are printed out to the file *_ion-den.txt
and then the ion density and ion potential are printed out every
1000 time steps.