Fixed a loop index in move dust loop
The loop that resets accDust2 for all of the dust particles durring
the first loop over the dust particles when moving the dust particles
in the time step in IonWake_000.cu. The code had been copied from
matlab with a starting index of 1. The starting index was changed
to 0 to mach C++.

An unused checkIonDustBounds_101 was removed, as boundsIon is cleared
before it is nexed used.

Fixed some formating.