Strong confinement for dust if abs(z)>0.8*HT_CYL, dust_dt=1e-4
Dust confinement changed to have parabolic confinement in z if the
absolute value of the dust z-position exceeds 0.8 * HT_CYL.
The dust time step has been hard coded to be 1e-4 seconds.  This saves
having to adjust the multiplier in the code
(dust_dt = N * 100 * TIME_STEP)
when the ion time step is changed in the parameter file.