Adjusted the dust charge for the dust-dust forces.
The dust charges to a potential relative to the background potential, which
here is set by Te (in eV) since the electrons are not included in the
simulation.  However, in the real dusty plasma experiment, the background
potential is essentially zero.  Hence the dust particles are overcharged
compared to experiment.  So this additional charge is subtracted from the dust
charge (by removing electrons) before the dust-dust forces are calculated