Average ion force on dust by N_IONDT_PER_DUSTDT.  Use adj_q for qdust*Ez
The force from ions on the dust was summed each ion time step, but needed
to be averaged before applying to the dust.  Since the dust is overcharged
due to the positive ion background, need to adjust the charge before
calculating forces on the dust. Note that the dust charge is not adjusted
when calculating the ion-dust interaction.  Should explore this.