sumDustIonacc_103 calculated with numBlocks = NUM_DUST
The total number of entries in accDustIon is NUM_ION*NUM_DUST.
Ordinarily, device processes loop over X = blocksPerGridIon blocks
which each have a block size of 1024. Instead, here we loop over
NUM_DUST blocks. Each dustBlock is responsible for summing the total
force on one dust grain. Each of the 1024 threads loads and sums
data for the X ion blocks.  The final reduction is based on the
threadID.