Changed radial confinement to be flat across cylinder, then parabolic.
This mimics the radial confinement in the azimuthal direction.  The
value for OMEGA2 is now called that in the parameter file, and the
value for RADIAL_CONFINE now has the same context as AXIAL_CONFINE
as it is a fraction to multiply by RAD_CYL. Thus dust particles which
have a radial position > RADIAL_CONFINE * RAD_CYL will have a steep
linear force exerted in the radial direction.