Averaged dust charge over 10 dust time steps.
The oscillations in the dust charge are quite large, and become
magnified when the dust charged is reduced by adj_q for calculating
the forces on the dust.  Averaging the charge over several dust time
steps comes closer to capturing the average charge over one dust step.
The oscillation of the charge actually occurs on a much faster time scale,
but by only using a fraction of the ion time steps which elapse over a
dust time step, the oscillations occur on the dust time scale, which is
unphysical.  Thus averaging over several dust time steps reduces the
effects of these oscillations.