Added params GAS_TYPE and BOX_CENTER; forces in lab sheath
GAS_TYPE = 1 for neon, 2 for argon
BOX_CENTER = height of center of simulation cylinder above lower electrode
Forces for the lab sheath include gravity, 5th order polynomial for
the sheath electric field (commented out).  Also commented out the
forces of the ions outside the simulation region acting on the dust,
as this force destabilizes the dust structure, pushing dust away from
the center of the simulation region.  Need to add options to turn on or off
various forces depending on the conditions.