Moved constCUDAvar.compare() to destructor
The constCUDAvar.compare() method is provided to check that
the value of the constCUDAcar hasn't been changed. Originally
this was called for each constCUDAvar at the end of the driver.
However this does not enforce constness checking. To reduce the
size of the driver and to enforce constness checking this call
has been moved to ~constCUDAvar. Now, if the constCUDAvar was changed
the destructor throes a ConstCUDAvarChanged. In the future the
destructor should be exception safe.