Added Doxygen compatibility
Doxygen is a program for automatic documentation generation. The
doxyfile specifies options for the documentation generation. To
generate the documentation run "doxygen" from the command line. It
is already isntaled on Kodiak. The documentation will be placed
in a directory labeled doc. Currently html, latex, and man pages
are generated.

The doc directory created by Doxygen has been added to the .gitignore
file as it trivially created.