Corrected initialization of accDust, added LASER_ON/OFF to params.
accDust[i].z was initialized with -= q/m*E, instead of = q/m*E.  In
some cases, the initial memory grabbed by malloc was good, but adding
the extra parameters to the parameter file caused malloc to grab bad
values.  LASER_ON and LASER_OFF added to the parameter file to specify the times to add an acceleration to the second dust particle.