Testing increasing the ion-dust collection radius to 3*dust radius.
In the Piel 2017 paper, the collection radius depends on the dust
potential and the ion speed.  At Mach 1, the collection radius is
approximately 3.5*dust_radius.  However, at Mach = 0.33, which is a
common number for me to use, Piel's formula results in a collection radius
of ~20*dust_radius, which is too big. Since I am accurately resolving
the position of the ions anyway, it shouldn't matter that I am using
a smaller radius.  But this may keep the ion-dust forces from "spiking"
when an ion passes very near a dust grain.