Moved location of sum_ionDust103 and zero_ionDust103
The ion forces on the dust are now summed after the ion
time step loop, before the MOVE_DUST check. This way, the
ion forces are summed whether or not the dust is moving.
The ion-dust forces are printed out after the end over the
MOVE_DUST section.  After the forces are printed out, then
the ion-dust forces are zeroed using zero_ionDust-103, just
before the end of the time step loop. Thus they are prepared
for the next set of ion-timestep loops.